Version 13 - success
Success in creating documentation, dialog, presenter, and models modules that compile without error. Only if a table column header uses a reserved name (e.g. "Date") does it not compile, which is to be expected. Possibly fix it in the future.

For some reason the documentation module is sporadic in writing the details. If processing is interrupted (i.e. using a breakpoint) at the location the details are assembled the details string is returned in full and the module gets the code. If it runs without interruption then the header and footer get in, but the details don't.